Add float processing

+++ qap17_9_1/utils.py
@@ -2,12 +2,12 @@ from extensions import UserInputException
 
 def validate(user_str, user_num):
   try:
-    items = [int(item) for item in user_str.split(' ')]
+    items = [float(item) for item in user_str.split(' ')]
   except ValueError:
     raise UserInputException(f'Невозможно обработать последовательность: {user_str}')
 
   try:
-    num = int(user_num)
+    num = float(user_num)
   except ValueError:
     raise UserInputException(f'Невозможно обработать число: {user_num}')
